HMB
In this paper, I tried to answer whether FDI has a considerable role in generating economic growth and vice-versa. Then, I provided economic forecasts for both real GDP and FDI for Turkey. And as a result I came to conclusion that the regression model and 5-year forecasts show that foreign direct investment has statistically no effect in generating real economic growth but potential positive effects for Turkey since it enables the transfer of knowledge and skills, increase the investment capacity, and create multiplier effect from the investment activity which further boost the aggregate demand and establishes more structural and persistent economic growth benefits.